+++ source/NVDAHelper/__init__.py
@@ -59,13 +59,23 @@ match sysconfig.get_platform():
 
 
 def __getattr__(name: str) -> Any:
+	def warnDeprecatedWithReplacement(deprecated: str, replacement: str):
+		warnings.warn(
+			f"NVDAHelper.{deprecated}is deprecated. Use {replacement} instead.",
+			DeprecationWarning,
+			stacklevel=3,
+		)
+
 	if NVDAState._allowDeprecatedAPI():
-		if name == "versionedLibPath":
-			warnings.warn(
-				"NVDAHelper.versionedLibPath is deprecated. Use NVDAHelper.versionedLibX86Path instead.",
-				DeprecationWarning,
-			)
-			return versionedLibX86Path
+		match name:
+			case "versionedLibPath":
+				warnDeprecatedWithReplacement(name, "NVDAHelper.versionedLibX86Path")
+				return versionedLibX86Path
+			case "generateBeep":
+				warnDeprecatedWithReplacement("generateBeep", "NVDAHelper.localLib.generateBeep")
+				return localLib.generateBeep
+			case _:
+				pass
 	raise AttributeError(f"Module {__name__!r} has no attribute {name!r}")
 
 
@@ -93,7 +103,6 @@ windll.kernel32.LoadLibraryExW.restype = HMODULE
 _remoteLib = None
 _remoteLoaderAMD64: "Optional[_RemoteLoader]" = None
 _remoteLoaderARM64: "Optional[_RemoteLoader]" = None
-generateBeep = None
 onSsmlMarkReached = None
 VBuf_getTextInRange = None
 lastLanguageID = None
@@ -798,7 +807,7 @@ class _RemoteLoader:
 
 def initialize() -> None:
 	global _remoteLib, _remoteLoaderAMD64, _remoteLoaderARM64
-	global generateBeep, onSsmlMarkReached, VBuf_getTextInRange
+	global onSsmlMarkReached, VBuf_getTextInRange
 	global lastLanguageID, lastLayoutString
 	hkl = c_ulong(windll.User32.GetKeyboardLayout(0)).value
 	lastLanguageID = winUser.LOWORD(hkl)
@@ -883,7 +892,7 @@ def initialize() -> None:
 
 def terminate():
 	global _remoteLib, _remoteLoaderAMD64, _remoteLoaderARM64
-	global generateBeep, VBuf_getTextInRange
+	global VBuf_getTextInRange
 	if not config.isAppX:
 		if not _remoteLib.uninstallIA2Support():
 			log.debugWarning("Error uninstalling IA2 support")
@@ -896,7 +905,6 @@ def terminate():
 		if _remoteLoaderARM64:
 			_remoteLoaderARM64.terminate()
 			_remoteLoaderARM64 = None
-	generateBeep = None
 	VBuf_getTextInRange = None
 	localLib.nvdaHelperLocal_terminate()